Functionality to display more than one route on a map
By adding a new source to the map ("map.addSource()") when the map
is loaded ("map.on()"), and subsequently drawing that route
("map.addLayer()"), we can draw more than one route on a single map.
That is, when a map is loaded at once, each route is added in order.

TODO:
- Call these routes from a database, rather than hardcode them.
- ...Build the database of routes!